refactor: refactor icons logo
