--- src/components/icons/index.ts
@@ -1,4 +1,0 @@
-export { Spinner } from "./spinner";
-export { Netflix } from "./netflix";
-export { Google } from "./google";
-export { Github } from "./github";
