+++ src/components/icons/logo.tsx
@@ -1,0 +1,104 @@
+interface LogoProps extends React.SVGProps<SVGSVGElement> {}
+
+export const Github = ({ ...props }: LogoProps) => {
+  return (
+    <svg
+      stroke="currentColor"
+      fill="currentColor"
+      strokeWidth="0"
+      viewBox="0 0 496 512"
+      height="1em"
+      width="1em"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M165.9 397.4c0 2-2.3 3.6-5.2 3.6-3.3.3-5.6-1.3-5.6-3.6 0-2 2.3-3.6 5.2-3.6 3-.3 5.6 1.3 5.6 3.6zm-31.1-4.5c-.7 2 1.3 4.3 4.3 4.9 2.6 1 5.6 0 6.2-2s-1.3-4.3-4.3-5.2c-2.6-.7-5.5.3-6.2 2.3zm44.2-1.7c-2.9.7-4.9 2.6-4.6 4.9.3 2 2.9 3.3 5.9 2.6 2.9-.7 4.9-2.6 4.6-4.6-.3-1.9-3-3.2-5.9-2.9zM244.8 8C106.1 8 0 113.3 0 252c0 110.9 69.8 205.8 169.5 239.2 12.8 2.3 17.3-5.6 17.3-12.1 0-6.2-.3-40.4-.3-61.4 0 0-70 15-84.7-29.8 0 0-11.4-29.1-27.8-36.6 0 0-22.9-15.7 1.6-15.4 0 0 24.9 2 38.6 25.8 21.9 38.6 58.6 27.5 72.9 20.9 2.3-16 8.8-27.1 16-33.7-55.9-6.2-112.3-14.3-112.3-110.5 0-27.5 7.6-41.3 23.6-58.9-2.6-6.5-11.1-33.3 2.6-67.9 20.9-6.5 69 27 69 27 20-5.6 41.5-8.5 62.8-8.5s42.8 2.9 62.8 8.5c0 0 48.1-33.6 69-27 13.7 34.7 5.2 61.4 2.6 67.9 16 17.7 25.8 31.5 25.8 58.9 0 96.5-58.9 104.2-114.8 110.5 9.2 7.9 17 22.9 17 46.4 0 33.7-.3 75.4-.3 83.6 0 6.5 4.6 14.4 17.3 12.1C428.2 457.8 496 362.9 496 252 496 113.3 383.5 8 244.8 8zM97.2 352.9c-1.3 1-1 3.3.7 5.2 1.6 1.6 3.9 2.3 5.2 1 1.3-1 1-3.3-.7-5.2-1.6-1.6-3.9-2.3-5.2-1zm-10.8-8.1c-.7 1.3.3 2.9 2.3 3.9 1.6 1 3.6.7 4.3-.7.7-1.3-.3-2.9-2.3-3.9-2-.6-3.6-.3-4.3.7zm32.4 35.6c-1.6 1.3-1 4.3 1.3 6.2 2.3 2.3 5.2 2.6 6.5 1 1.3-1.3.7-4.3-1.3-6.2-2.2-2.3-5.2-2.6-6.5-1zm-11.4-14.7c-1.6 1-1.6 3.6 0 5.9 1.6 2.3 4.3 3.3 5.6 2.3 1.6-1.3 1.6-3.9 0-6.2-1.4-2.3-4-3.3-5.6-2z"></path>
+    </svg>
+  );
+};
+
+export const Google = ({ ...props }: LogoProps) => {
+  return (
+    <svg
+      stroke="currentColor"
+      fill="currentColor"
+      strokeWidth="0"
+      version="1.1"
+      x="0px"
+      y="0px"
+      viewBox="0 0 48 48"
+      enableBackground="new 0 0 48 48"
+      height="1em"
+      width="1em"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        fill="#FFC107"
+        d="M43.611,20.083H42V20H24v8h11.303c-1.649,4.657-6.08,8-11.303,8c-6.627,0-12-5.373-12-12c0-6.627,5.373-12,12-12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C12.955,4,4,12.955,4,24	c0,11.045,8.955,20,20,20c11.045,0,20-8.955,20-20C44,22.659,43.862,21.35,43.611,20.083z"
+      ></path>
+      <path
+        fill="#FF3D00"
+        d="M6.306,14.691l6.571,4.819C14.655,15.108,18.961,12,24,12c3.059,0,5.842,1.154,7.961,3.039l5.657-5.657C34.046,6.053,29.268,4,24,4C16.318,4,9.656,8.337,6.306,14.691z"
+      ></path>
+      <path
+        fill="#4CAF50"
+        d="M24,44c5.166,0,9.86-1.977,13.409-5.192l-6.19-5.238C29.211,35.091,26.715,36,24,36c-5.202,0-9.619-3.317-11.283-7.946l-6.522,5.025C9.505,39.556,16.227,44,24,44z"
+      ></path>
+      <path
+        fill="#1976D2"
+        d="M43.611,20.083H42V20H24v8h11.303c-0.792,2.237-2.231,4.166-4.087,5.571c0.001-0.001,0.002-0.001,0.003-0.002l6.19,5.238C36.971,39.205,44,34,44,24C44,22.659,43.862,21.35,43.611,20.083z"
+      ></path>
+    </svg>
+  );
+};
+
+export const Netflix = ({ ...props }: LogoProps) => {
+  return (
+    <svg
+      width={100}
+      height="1.5rem"
+      viewBox="0 0 500 150"
+      fill="#E50914"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M0 9.20296e-05H26.4382L46.9374 78.5626H47.3206V9.20296e-05H66.0955V65.6251V107.297V123.163L43.9995 125.668L38.1247 107.297L31.8025 83.3438L19.1581 35.4376H18.775V129.084L0 132.102V9.20296e-05Z"
+        fill="#E50914"
+      />
+      <path
+        d="M173.194 18.7501H151.162V9.20296e-05H216.299V18.7501H194.268V114.958L173.194 115.667V18.7501Z"
+        fill="#E50914"
+      />
+      <path
+        d="M226.503 9.20296e-05H282.253V18.7501H247.577V32.8126V38.1947H274.781V56.9447H247.577V114.132L226.503 114.294V48.5072V32.8126V9.20296e-05Z"
+        fill="#E50914"
+      />
+      <path
+        d="M292.359 9.20296e-05H313.433V74.0626V95.6242H348.109V115.751C348.109 115.751 331.12 115.331 320.234 115.062C309.348 114.794 292.359 114.374 292.359 114.374V74.0626V9.20296e-05Z"
+        fill="#E50914"
+      />
+      <path
+        d="M390 117.84C376.9 117.001 355.934 116.062 355.934 116.062V100.125C355.934 100.125 355.934 40.625 355.934 30.125C355.934 19.625 355.934 9.20296e-05 355.934 9.20296e-05H387.928H419.922V15.5626V31.1251V100.125C419.922 110.625 419.922 120.09 419.922 120.09C419.922 120.09 401.889 118.602 390 117.84ZM387.928 98.125H398.848V18.7501H387.928H377.008V98.125H387.928Z"
+        fill="#E50914"
+      />
+      <path
+        d="M449.764 64.1251L426.582 9.20296e-05H448.806L462.983 42.3751H463.366L477.926 9.20296e-05H497.851L474.669 64.1251L499 131.25L477.926 127.198L461.45 85.5001H461.067L448.5 123.036L430.5 121.077L449.764 64.1251Z"
+        fill="#E50914"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M141.488 117.052C141.488 117.052 121.96 118.18 109.498 119.083C119.965 118.321 130.643 117.643 141.488 117.052ZM108.568 119.151C95.9721 120.079 77.4995 121.737 77.4995 121.737C87.6053 120.791 97.9775 119.927 108.568 119.151Z"
+        fill="#E50914"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M77.4995 100.125V121.737C87.6053 120.791 97.9775 119.927 108.568 119.151C108.719 119.139 108.87 119.128 109.019 119.117C109.178 119.106 109.336 119.094 109.494 119.083L109.498 119.083C119.965 118.321 130.643 117.643 141.488 117.052V100.125V31.125V15.5625V0L109.494 9.20296e-05L77.4995 0V30.1249V100.125ZM120.414 98.1249H109.494H98.5735V18.75H109.494H120.414V98.1249Z"
+        fill="#E50914"
+      />
+    </svg>
+  );
+};
